% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harlem.R
\docType{package}
\name{harlem}
\alias{harlem}
\title{harlem}
\description{
A collection of color palettes inspired by iconic works from Harlem Renaissance artists.
}
\details{
A collection of color palettes inspired by Harlem Renaissance artists and their pieces of art.
}
