% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{harlem_palette}
\alias{harlem_palette}
\title{Harlem Renaissance Palette Generator}
\usage{
harlem_palette(
  name = "LawrenceStreetScene",
  n = NULL,
  type = c("discrete", "continuous")
)
}
\arguments{
\item{name}{Character string. Name of the desired palette. Available choices are included in the package documentation.}

\item{n}{Integer. Number of colors desired. If omitted, uses all available colors from the chosen palette.}

\item{type}{Character string. Either "continuous" or "discrete". Use "continuous" to interpolate between colors.}
}
\value{
A vector of colors.
}
\description{
Generates color palettes inspired by artworks from the Harlem Renaissance era.
}
\examples{
harlem_palette("LawrenceStreetScene")
harlem_palette("BeardenConjurWoman", 5)
harlem_palette("MotleyBlues", type = "continuous")
}
\keyword{colors}
