% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{harlem_palettes}
\alias{harlem_palettes}
\title{A list of Harlem Renaissance-inspired palettes}
\format{
An object of class \code{list} of length 10.
}
\usage{
harlem_palettes
}
\description{
A list of Harlem Renaissance-inspired palettes
}
\keyword{datasets}
