% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{print_palette}
\alias{print_palette}
\title{Print the Palette Colors}
\usage{
print_palette(x, ...)
}
\arguments{
\item{x}{A color palette represented as a vector of color hex codes.}

\item{...}{Additional arguments to be passed to the underlying \code{image} and \code{text} functions.}
}
\value{
A plot displaying the color palette.
}
\description{
This function displays the color palette visually as an image plot.
The palette name, if it exists, will be displayed as a label on the plot.
}
\examples{
my_palette <- c("#D7AA50", "#C13E20", "#3E5D33", "#E6E6E6")
attr(my_palette, "name") <- "Example Palette"
print_palette(my_palette)
}
